ai_api_intg/requirements.txt
groq
huggingface_hub
cohere
requests
python-dotenv